Add files via upload
In this project, we enhanced the Persons1 table by adding a DOB column, created a function to calculate age, retrieved ages using this function, and demonstrated how to perform inner, left, and right joins between the Persons1 and Count1 tables. These operations are essential for managing and querying relational database tables effectively.